Move string literal tests to separate file
