+++ src/test/resources/samples/modifiers.ts
@@ -1,7 +1,6 @@
 declare module monaco {
 
   const id: string;
-  export type BuiltinTheme = 'vs' | 'vs-dark' | 'hc-black';
 
   export class Emitter<T> {
       constructor();
@@ -31,11 +30,4 @@ declare module monaco {
       readonly path: string;
   }
 
-  export interface IEditorOptions {
-      ariaLabel?: string;
-      rulers?: number[];
-      selectionClipboard?: boolean;
-      lineNumbers?: 'on' | 'off' | 'relative' | ((lineNumber: number) => string);
-      readable?: 'yes' | boolean | 'restricted';
-  }
 }
